+++ jmd/__main__.py
@@ -4,4 +4,4 @@
 from ._cli import main
 
 if __name__ == "__main__":
-    main()
+    raise SystemExit(main())